updated the survey form and the database
Updated the survey form to take in multiple instruments, corresponding skill level and genre ratings.

Also updated the database to which now containt 4 table: profile which stores most of the profile info. Instruments table which maps from profile to instruments played. skill which maps from instrument played to skill level and genre which maps a profile to genre preferences.